# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kabianga ARG Portal API",
        "version": "1.0.0",
        "description": "Research grants management API"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "paths": {
...
        "/api/v1/settings": {
            "get": {
                "tags": ["Settings"],
                "summary": "Get application settings",
                "security": [{"bearerAuth": []}],
                "responses": {
                    "200": {
                        "description": "Application settings"
                    }
                }
            },
            "put": {
                "tags": ["Settings"],
                "summary": "Update application settings",
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "setting_key": {"type": "string"},
                                    "setting_value": {"type": "string"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated"
                    }
                }
            }
        },
...
        "/api/v1/proposals/search": {
            "get": {
                "tags": ["Proposals"],
                "summary": "Search proposals",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {"type": "string"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                }
            }
        },
...
        "/api/v1/publications/search": {
            "get": {
                "tags": ["Publications"],
                "summary": "Search publications",
                "security": [{"bearerAuth": []}],
                "responses": {"200": {"description": "Search results"}}
            }
        },
...
        "/api/v1/research-design/search": {
            "get": {
                "tags": ["Research Design"],
                "summary": "Search research designs",
                "security": [{"bearerAuth": []}],
                "responses": {"200": {"description": "Search results"}}
            }
        },
...
    }
}